redid particles for restore

+++ src/main/java/com/crowsofwar/avatar/common/bending/water/WaterParticleSpawner.java
@@ -29,7 +29,7 @@ public class WaterParticleSpawner extends TickHandler {
 			maxDuration = 60;
 		}
 		int duration = data.getTickHandlerDuration(this);
-		double radius =  (maxDuration - duration) / 10;
+		double radius =  ((float) maxDuration - duration) / 10;
